test(sqlx): wire per-type <T>_ord_ope smoke modules into catalog-coverage
The test:matrix:catalog-coverage gate requires every catalog (type, domain)
pair to have matrix wiring: either macro-emitted scalars::<t>::matrix_* names
or a dedicated top-level <t>_<seg>::* module (the text_match mechanism). The
single ope_smoke module satisfied neither pattern, so the gate reported every
ord_ope domain as MISSING.

Restructure the ope suite: ope_support carries the shared literal-payload
builder (hm + op — CHECKs assert key presence, so one builder serves the
integer [Ope] domains and text's [Hm, Ope]) and an ope_ord_smoke! macro
stamping routing-agnostic ordering/equality/CHECK tests; nine top-level
<t>_ord_ope modules invoke it (int4 keeps the deeper prefix/blocker/ORDER
BY/aggregate tests; text_ord_ope keeps the hm-routed equality contract).

All 32 tests pass against a live Postgres 17 with the built v3 installer;
test:matrix:catalog-coverage, test:matrix:inventory and the other creds-free
gates are green.
